Refactor QRT implementation files to: rt (arrays and tuples); qis (base and msft compatible)
